Move role_ids_to_names to the utils resource from viewbinds
