+++ src/resources/utils.py
@@ -1,6 +1,9 @@
 from enum import Enum
 import aiohttp, asyncio
 
+import bloxlink
+import hikari
+
 from requests.utils import requote_uri
 from .secrets import PROXY_URL
 from .exceptions import RobloxAPIError, RobloxDown, RobloxNotFound
@@ -139,3 +142,26 @@ async def fetch(
     except asyncio.TimeoutError:
         print(f"URL {old_url} timed out", flush=True)
         raise RobloxDown()
+
+async def role_ids_to_names(guild_id: int, roles: list) -> str:
+    # TODO: utilize in-dev cache logic to get role data (and by extension the names)
+    # for now, I will just always query for guild data. (very much a not friendly request pattern)
+
+    guild: hikari.guilds.RESTGuild = await bloxlink.rest.fetch_guild(guild_id)
+    guild_roles = guild.roles
+
+    output_list = []
+
+    for role in roles:
+        output_list.append(
+            guild_roles.get(hikari.Snowflake(role)).name
+            if guild_roles.get(hikari.Snowflake(role)) is not None
+            else "(Deleted Role)"
+        )
+
+    if len(output_list) == 0:
+        return ""
+    elif len(output_list) > 1:
+        return ", ".join(output_list)
+    else:
+        return output_list[0]
